requirements.txt
apscheduler
aiogram
python-dotenv
tortoise-orm[aiomysql]
redis[hiredis]